Fix styles and Rename filenames
