+++ src/components/check-box/check-box.scss
