+++ src/layouts/auth-layout/auth-layout.jsx
@@ -1,12 +1,13 @@
 import PropTypes from 'prop-types';
 import React from 'react';
+import Logo from '../../svgs/logo.svg';
 
 export default function AuthLayout({ children }) {
   return (
     <div className="auth-layout">
       <div className="auth-header">
         <a href="/" className="auth-logo">
-          <img src="/images/logo.svg" alt="Home" className="auth-logo__img" />
+          <Logo className="auth-logo__img" alt="Home" />
         </a>
         <p className="auth-greeting">
           Join the community
