+++ src/styles/index.scss
@@ -6,13 +6,15 @@
 @import './elements/index';
 @import './objects/index';
 @import '~@/components/app-header/app-header';
+@import '~@/components/app-logo/app-logo';
 @import '~@/components/app-nav/app-nav';
 @import '~@/components/button/button';
-@import '~@/components/checkbox/checkbox';
+@import '~@/components/check-box/check-box';
+@import '~@/components/drop-down/drop-down';
 @import '~@/components/form-field/form-field';
 @import '~@/components/text-field/text-field';
 @import '~@/layouts/auth-layout/auth-layout';
 @import '~@/layouts/default-layout/default-layout';
 @import './pages/index';
-@import '~@/pages/forum-list/forum-list';
+@import '~@/pages/forum/list/forum-list';
 @import './utils/index';
