+++ src/styles/pages/_index.scss
@@ -1,2 +1,1 @@
 @import './auth';
-@import './container';
